Added pagination

+++ movies/models.py
@@ -7,7 +7,7 @@ from django.db import models
 class Movie(models.Model):
     name = models.CharField(max_length=255)
     year = models.IntegerField()
-    imdb_rating = models.IntegerField()
+    imdb_rating = models.FloatField()
     metascore = models.IntegerField()
     votes = models.IntegerField()
     description = models.CharField(max_length=2000)
